Test Longitudinal Measurement Invariance

Test Longitudinal Measurement Invariance

Usage

Invariance(
  data,
  time_points,
  factor_loadings,
  covariances = FALSE,
  model_add_configural = NULL,
  model_add_weak = NULL,
  model_add_strong = NULL,
  model_add_strict = NULL,
  ...
)

Arguments

data

Dataframe. The function assumes that the data is in the wide format and the variables are named as follows: paste0("y", time_point, "_", item_number). For example, for the item 1 from the first time point, the variable name should be y1_1.

time_points

Numeric vector of discrete time points.

factor_loadings

List with length equal to the number of factors. Each element of the list is the item number of items for the specific factor.

covariances

Logical. If covariance = TRUE, model the covariances of the measurement error.

model_add_configural

Additional specification added to the lavaan model syntax for the configural invariance model.

model_add_weak

Additional specification added to the lavaan model syntax for the weak invariance model.

model_add_strong

Additional specification added to the lavaan model syntax for the strong invariance model.

model_add_strict

Additional specification added to the lavaan model syntax for the strict invariance model.

...

Additional arguments to pass to lavaan::cfa().

Value

Returns an object of class longmi which is a list with the following elements:

call

Function call.

args

List of function arguments.

fit

Fitted models.

fun

Function used ("Invariance").

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

data("osbornesudick1972", package = "longMI")
mi <- Invariance(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
print(mi)
#> Call:
#> Invariance(data = osbornesudick1972, time_points = c(1, 6), factor_loadings = list(c(1, 
#>     2, 3, 4)))
#> 
#>               chisq df pvalue    cfi    tli  rmsea   srmr      aic      bic
#> configural  25.9682 19 0.1311 0.9915 0.9875 0.0424 0.0306 11252.23 11335.18
#> weak        41.8973 22 0.0064 0.9757 0.9691 0.0666 0.0720 11262.16 11335.16
#> strong      53.7228 25 0.0007 0.9650 0.9608 0.0750 0.0859 11267.98 11331.03
#> strict     134.5591 29 0.0000 0.8712 0.8757 0.1336 0.1328 11340.82 11390.59
summary(mi)
#> Call:
#> Invariance(data = osbornesudick1972, time_points = c(1, 6), factor_loadings = list(c(1, 
#>     2, 3, 4)))
#> 
#> Fit Measures
#>               chisq df pvalue    cfi    tli  rmsea   srmr      aic      bic
#> configural  25.9682 19 0.1311 0.9915 0.9875 0.0424 0.0306 11252.23 11335.18
#> weak        41.8973 22 0.0064 0.9757 0.9691 0.0666 0.0720 11262.16 11335.16
#> strong      53.7228 25 0.0007 0.9650 0.9608 0.0750 0.0859 11267.98 11331.03
#> strict     134.5591 29 0.0000 0.8712 0.8757 0.1336 0.1328 11340.82 11390.59
#> 
#> 
#> 
#> CONFIGURAL INVARIANCE MODEL
#> 
#> lavaan 0.6.16 ended normally after 75 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        27
#>   Number of equality constraints                     2
#> 
#>   Number of observations                           204
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                25.968
#>   Degrees of freedom                                19
#>   P-value (Chi-square)                           0.131
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 =~                                             
#>     y1_1     (l11)    4.451    0.396   11.244    0.000
#>     y1_2              6.850    0.637   10.750    0.000
#>     y1_3              4.590    0.515    8.918    0.000
#>     y1_4              5.039    0.393   12.809    0.000
#>   f1t6 =~                                             
#>     y6_1     (l11)    4.451    0.396   11.244    0.000
#>     y6_2              4.006    0.485    8.261    0.000
#>     y6_3              4.551    0.546    8.342    0.000
#>     y6_4              4.102    0.448    9.150    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 ~~                                             
#>     f1t6              1.837    0.211    8.703    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1      (i1)   19.776    0.427   46.273    0.000
#>    .y1_2             21.797    0.680   32.036    0.000
#>    .y1_3             14.903    0.528   28.223    0.000
#>    .y1_4             20.396    0.439   46.416    0.000
#>    .y6_1      (i1)   19.776    0.427   46.273    0.000
#>    .y6_2             19.317    2.281    8.467    0.000
#>    .y6_3             11.922    2.538    4.697    0.000
#>    .y6_4             17.970    1.815    9.903    0.000
#>     f1t1              0.000                           
#>     f1t6              6.455    0.601   10.743    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1             17.448    2.186    7.981    0.000
#>    .y1_2             47.511    5.748    8.266    0.000
#>    .y1_3             35.810    3.969    9.022    0.000
#>    .y1_4             13.999    2.085    6.712    0.000
#>    .y6_1             47.096    6.305    7.470    0.000
#>    .y6_2             73.850    8.395    8.797    0.000
#>    .y6_3             88.920   10.222    8.699    0.000
#>    .y6_4             23.267    4.076    5.709    0.000
#>     f1t1              1.000                           
#>     f1t6              5.834    1.159    5.035    0.000
#> 
#> 
#> 
#> WEAK INVARIANCE MODEL
#> 
#> lavaan 0.6.16 ended normally after 54 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        27
#>   Number of equality constraints                     5
#> 
#>   Number of observations                           204
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                41.897
#>   Degrees of freedom                                22
#>   P-value (Chi-square)                           0.006
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 =~                                             
#>     y1_1     (l11)    4.933    0.342   14.420    0.000
#>     y1_2     (l12)    5.172    0.413   12.534    0.000
#>     y1_3     (l13)    5.072    0.398   12.733    0.000
#>     y1_4     (l14)    4.865    0.330   14.733    0.000
#>   f1t6 =~                                             
#>     y6_1     (l11)    4.933    0.342   14.420    0.000
#>     y6_2     (l12)    5.172    0.413   12.534    0.000
#>     y6_3     (l13)    5.072    0.398   12.733    0.000
#>     y6_4     (l14)    4.865    0.330   14.733    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 ~~                                             
#>     f1t6              1.558    0.136   11.482    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1      (i1)   19.776    0.445   44.430    0.000
#>    .y1_2             21.797    0.629   34.633    0.000
#>    .y1_3             14.903    0.544   27.411    0.000
#>    .y1_4             20.396    0.436   46.803    0.000
#>    .y6_1      (i1)   19.776    0.445   44.430    0.000
#>    .y6_2             15.049    2.223    6.770    0.000
#>    .y6_3             11.756    2.238    5.253    0.000
#>    .y6_4             16.111    1.699    9.484    0.000
#>     f1t1              0.000                           
#>     f1t6              5.824    0.432   13.476    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1             16.079    2.165    7.427    0.000
#>    .y1_2             54.055    5.831    9.270    0.000
#>    .y1_3             34.578    3.922    8.817    0.000
#>    .y1_4             15.075    2.051    7.350    0.000
#>    .y6_1             49.748    6.242    7.970    0.000
#>    .y6_2             72.254    8.498    8.503    0.000
#>    .y6_3             91.610   10.272    8.919    0.000
#>    .y6_4             22.022    3.989    5.521    0.000
#>     f1t1              1.000                           
#>     f1t6              4.240    0.539    7.867    0.000
#> 
#> 
#> 
#> STRONG INVARIANCE MODEL
#> 
#> lavaan 0.6.16 ended normally after 52 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        27
#>   Number of equality constraints                     8
#> 
#>   Number of observations                           204
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                53.723
#>   Degrees of freedom                                25
#>   P-value (Chi-square)                           0.001
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 =~                                             
#>     y1_1     (l11)    5.270    0.333   15.813    0.000
#>     y1_2     (l12)    4.525    0.308   14.681    0.000
#>     y1_3     (l13)    4.960    0.328   15.120    0.000
#>     y1_4     (l14)    4.547    0.289   15.707    0.000
#>   f1t6 =~                                             
#>     y6_1     (l11)    5.270    0.333   15.813    0.000
#>     y6_2     (l12)    4.525    0.308   14.681    0.000
#>     y6_3     (l13)    4.960    0.328   15.120    0.000
#>     y6_4     (l14)    4.547    0.289   15.707    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 ~~                                             
#>     f1t6              1.608    0.143   11.281    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1      (i1)   19.929    0.456   43.727    0.000
#>    .y1_2      (i2)   21.459    0.598   35.868    0.000
#>    .y1_3      (i3)   14.882    0.529   28.153    0.000
#>    .y1_4      (i4)   20.311    0.421   48.237    0.000
#>    .y6_1      (i1)   19.929    0.456   43.727    0.000
#>    .y6_2      (i2)   21.459    0.598   35.868    0.000
#>    .y6_3      (i3)   14.882    0.529   28.153    0.000
#>    .y6_4      (i4)   20.311    0.421   48.237    0.000
#>     f1t1              0.000                           
#>     f1t6              5.337    0.353   15.109    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1             15.124    2.193    6.895    0.000
#>    .y1_2             57.819    6.078    9.512    0.000
#>    .y1_3             34.209    3.850    8.885    0.000
#>    .y1_4             16.339    2.062    7.925    0.000
#>    .y6_1             45.360    6.182    7.337    0.000
#>    .y6_2             74.229    8.350    8.890    0.000
#>    .y6_3             89.572   10.064    8.900    0.000
#>    .y6_4             24.586    3.878    6.339    0.000
#>     f1t1              1.000                           
#>     f1t6              4.557    0.584    7.806    0.000
#> 
#> 
#> 
#> STRICT INVARIANCE MODEL
#> 
#> lavaan 0.6.16 ended normally after 57 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        27
#>   Number of equality constraints                    12
#> 
#>   Number of observations                           204
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               134.559
#>   Degrees of freedom                                29
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 =~                                             
#>     y1_1     (l11)    5.083    0.364   13.968    0.000
#>     y1_2     (l12)    4.309    0.325   13.248    0.000
#>     y1_3     (l13)    4.785    0.354   13.521    0.000
#>     y1_4     (l14)    4.358    0.312   13.971    0.000
#>   f1t6 =~                                             
#>     y6_1     (l11)    5.083    0.364   13.968    0.000
#>     y6_2     (l12)    4.309    0.325   13.248    0.000
#>     y6_3     (l13)    4.785    0.354   13.521    0.000
#>     y6_4     (l14)    4.358    0.312   13.971    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 ~~                                             
#>     f1t6              1.812    0.168   10.790    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1      (i1)   20.019    0.506   39.547    0.000
#>    .y1_2      (i2)   21.513    0.621   34.617    0.000
#>    .y1_3      (i3)   14.805    0.617   24.001    0.000
#>    .y1_4      (i4)   20.313    0.433   46.903    0.000
#>    .y6_1      (i1)   20.019    0.506   39.547    0.000
#>    .y6_2      (i2)   21.513    0.621   34.617    0.000
#>    .y6_3      (i3)   14.805    0.617   24.001    0.000
#>    .y6_4      (i4)   20.313    0.433   46.903    0.000
#>     f1t1              0.000                           
#>     f1t6              5.557    0.411   13.526    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1      (u1)   28.657    2.747   10.433    0.000
#>    .y1_2      (u2)   68.013    5.170   13.155    0.000
#>    .y1_3      (u3)   61.387    4.821   12.734    0.000
#>    .y1_4      (u4)   20.878    2.008   10.398    0.000
#>    .y6_1      (u1)   28.657    2.747   10.433    0.000
#>    .y6_2      (u2)   68.013    5.170   13.155    0.000
#>    .y6_3      (u3)   61.387    4.821   12.734    0.000
#>    .y6_4      (u4)   20.878    2.008   10.398    0.000
#>     f1t1              1.000                           
#>     f1t6              5.056    0.685    7.378    0.000
#>