Strict Invariance Model

Strict Invariance Model

Usage

Strict(
  data,
  time_points,
  factor_loadings,
  covariances = FALSE,
  model_add = NULL,
  ...
)

Arguments

data

Dataframe. The function assumes that the data is in the wide format and the variables are named as follows: paste0("y", time_point, "_", item_number). For example, for the item 1 from the first time point, the variable name should be y1_1.

time_points

Numeric vector of discrete time points.

factor_loadings

List with length equal to the number of factors. Each element of the list is the item number of items for the specific factor.

covariances

Logical. If covariance = TRUE, model the covariances of the measurement error.

model_add

Additional specification added to the lavaan model syntax.

...

Additional arguments to pass to lavaan::cfa().

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

data("osbornesudick1972", package = "longMI")
strict_fit <- Strict(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
library(lavaan)
summary(strict_fit)
#> lavaan 0.6.16 ended normally after 57 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        27
#>   Number of equality constraints                    12
#> 
#>   Number of observations                           204
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                               134.559
#>   Degrees of freedom                                29
#>   P-value (Chi-square)                           0.000
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   t1f1 =~                                             
#>     y1_1     (l11)    5.083    0.364   13.968    0.000
#>     y1_2     (l12)    4.309    0.325   13.248    0.000
#>     y1_3     (l13)    4.785    0.354   13.521    0.000
#>     y1_4     (l14)    4.358    0.312   13.971    0.000
#>   t6f1 =~                                             
#>     y6_1     (l11)    5.083    0.364   13.968    0.000
#>     y6_2     (l12)    4.309    0.325   13.248    0.000
#>     y6_3     (l13)    4.785    0.354   13.521    0.000
#>     y6_4     (l14)    4.358    0.312   13.971    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   t1f1 ~~                                             
#>     t6f1              1.812    0.168   10.790    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1      (i1)   20.019    0.506   39.547    0.000
#>    .y1_2      (i2)   21.513    0.621   34.617    0.000
#>    .y1_3      (i3)   14.805    0.617   24.001    0.000
#>    .y1_4      (i4)   20.313    0.433   46.903    0.000
#>    .y6_1      (i1)   20.019    0.506   39.547    0.000
#>    .y6_2      (i2)   21.513    0.621   34.617    0.000
#>    .y6_3      (i3)   14.805    0.617   24.001    0.000
#>    .y6_4      (i4)   20.313    0.433   46.903    0.000
#>     t1f1              0.000                           
#>     t6f1              5.557    0.411   13.526    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1      (u1)   28.657    2.747   10.433    0.000
#>    .y1_2      (u2)   68.013    5.170   13.155    0.000
#>    .y1_3      (u3)   61.387    4.821   12.734    0.000
#>    .y1_4      (u4)   20.878    2.008   10.398    0.000
#>    .y6_1      (u1)   28.657    2.747   10.433    0.000
#>    .y6_2      (u2)   68.013    5.170   13.155    0.000
#>    .y6_3      (u3)   61.387    4.821   12.734    0.000
#>    .y6_4      (u4)   20.878    2.008   10.398    0.000
#>     t1f1              1.000                           
#>     t6f1              5.056    0.685    7.378    0.000
#>