Strong Invariance Model

Strong Invariance Model

Usage

Strong(
  data,
  time_points,
  factor_loadings,
  covariances = FALSE,
  model_add = NULL,
  ...
)

Arguments

data

Dataframe. The function assumes that the data is in the wide format and the variables are named as follows: paste0("y", time_point, "_", item_number). For example, for the item 1 from the first time point, the variable name should be y1_1.

time_points

Numeric vector of discrete time points.

factor_loadings

List with length equal to the number of factors. Each element of the list is the item number of items for the specific factor.

covariances

Logical. If covariance = TRUE, model the covariances of the measurement error.

model_add

Additional specification added to the lavaan model syntax.

...

Additional arguments to pass to lavaan::cfa().

Value

Returns a fitted lavaan object.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

data("osbornesudick1972", package = "longMI")
strong_fit <- Strong(
  data = osbornesudick1972,
  time_points = c(1, 6),
  factor_loadings = list(
    c(1, 2, 3, 4)
  )
)
library(lavaan)
summary(strong_fit)
#> lavaan 0.6.16 ended normally after 52 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        27
#>   Number of equality constraints                     8
#> 
#>   Number of observations                           204
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                53.723
#>   Degrees of freedom                                25
#>   P-value (Chi-square)                           0.001
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 =~                                             
#>     y1_1     (l11)    5.270    0.333   15.813    0.000
#>     y1_2     (l12)    4.525    0.308   14.681    0.000
#>     y1_3     (l13)    4.960    0.328   15.120    0.000
#>     y1_4     (l14)    4.547    0.289   15.707    0.000
#>   f1t6 =~                                             
#>     y6_1     (l11)    5.270    0.333   15.813    0.000
#>     y6_2     (l12)    4.525    0.308   14.681    0.000
#>     y6_3     (l13)    4.960    0.328   15.120    0.000
#>     y6_4     (l14)    4.547    0.289   15.707    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1t1 ~~                                             
#>     f1t6              1.608    0.143   11.281    0.000
#> 
#> Intercepts:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1      (i1)   19.929    0.456   43.727    0.000
#>    .y1_2      (i2)   21.459    0.598   35.868    0.000
#>    .y1_3      (i3)   14.882    0.529   28.153    0.000
#>    .y1_4      (i4)   20.311    0.421   48.237    0.000
#>    .y6_1      (i1)   19.929    0.456   43.727    0.000
#>    .y6_2      (i2)   21.459    0.598   35.868    0.000
#>    .y6_3      (i3)   14.882    0.529   28.153    0.000
#>    .y6_4      (i4)   20.311    0.421   48.237    0.000
#>     f1t1              0.000                           
#>     f1t6              5.337    0.353   15.109    0.000
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .y1_1             15.124    2.193    6.895    0.000
#>    .y1_2             57.819    6.078    9.512    0.000
#>    .y1_3             34.209    3.850    8.885    0.000
#>    .y1_4             16.339    2.062    7.925    0.000
#>    .y6_1             45.360    6.182    7.337    0.000
#>    .y6_2             74.229    8.350    8.890    0.000
#>    .y6_3             89.572   10.064    8.900    0.000
#>    .y6_4             24.586    3.878    6.339    0.000
#>     f1t1              1.000                           
#>     f1t6              4.557    0.584    7.806    0.000
#>